pyproject.toml
[tool.poetry]
name = "source"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
ccxt = "^4.3.65"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
ratelimit = "^2.2.1"
ibis-framework = {extras = ["duckdb"], version = "^9.5.0"}
streamlit = "^1.38.0"
plotly = "^5.24.1"
watchdog = "^4.0.2"
networkx = "^3.3"
nest-asyncio = "^1.6.0"
dash = "^2.18.1"
dash-bootstrap-components = "^1.6.0"
tabulate = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
